Split asyncio.coroutine tests and set max_pyver 3.10 for them
iterable_context_py3 includes some checks that we don't emit
errors for asyncio.coroutine, but asyncio.coroutine has been
removed in Python 3.11, so we need to set a max version of
3.10 for these tests.
