+++ tests/functional/i/iterable_context_asyncio.rc
@@ -1,0 +1,2 @@
+[testoptions]
+max_pyver=3.10
