+++ tests/functional/i/iterable_context_py3.py
@@ -16,40 +16,3 @@ for i in SomeClass:
     print(i)
 for i in SomeClass():  # [not-an-iterable]
     print(i)
-
-
-import asyncio
-
-
-@asyncio.coroutine
-def coroutine_function_return_none():
-    return
-
-
-@asyncio.coroutine
-def coroutine_function_return_object():
-    return 12
-
-
-@asyncio.coroutine
-def coroutine_function_return_future():
-    return asyncio.Future()
-
-
-@asyncio.coroutine
-def coroutine_function_pass():
-    pass
-
-
-@asyncio.coroutine
-def coroutine_generator():
-    yield
-
-
-@asyncio.coroutine
-def main():
-    yield from coroutine_function_return_none()
-    yield from coroutine_function_return_object()
-    yield from coroutine_function_return_future()
-    yield from coroutine_function_pass()
-    yield from coroutine_generator()
